fix(git): account for os when resolving ssh agent
