+++ internal/giturl/giturl.go
@@ -2,10 +2,7 @@ package giturl
 
 import (
 	"fmt"
-	"os"
-	"os/exec"
 	"strings"
-	"syscall"
 
 	"github.com/dargstack/dargstack/v4/internal/logger"
 )
@@ -50,29 +47,6 @@ func RepoNameFromURL(url string) string {
 	}
 	name = strings.TrimSuffix(name, ".git")
 	return name
-}
-
-// SSHAgentAvailable returns true if an SSH agent is running and accessible.
-func SSHAgentAvailable() bool {
-	sock := os.Getenv("SSH_AUTH_SOCK")
-	if sock == "" {
-		return false
-	}
-
-	info, err := os.Stat(sock)
-	if err != nil {
-		return false
-	}
-	if info.Mode()&os.ModeSocket == 0 {
-		return false
-	}
-
-	cmd := exec.Command("ssh-add", "-l")
-	cmd.Env = append(os.Environ(), "SSH_AUTH_SOCK="+sock)
-	cmd.SysProcAttr = &syscall.SysProcAttr{
-		Setpgid: true,
-	}
-	return cmd.Run() == nil
 }
 
 const (
