+++ internal/giturl/giturl_test.go
@@ -86,13 +86,6 @@ func TestRepoNameFromURL(t *testing.T) {
 	}
 }
 
-func TestSSHAgentAvailable_NoEnvVar(t *testing.T) {
-	t.Setenv("SSH_AUTH_SOCK", "")
-	if SSHAgentAvailable() {
-		t.Error("SSHAgentAvailable() should be false when SSH_AUTH_SOCK is not set")
-	}
-}
-
 func TestExtractFromService(t *testing.T) {
 	tests := []struct {
 		name          string
