+++ internal/giturl/giturl_unix_test.go
@@ -1,0 +1,12 @@
+//go:build !windows
+
+package giturl
+
+import "testing"
+
+func TestSSHAgentAvailable_NoEnvVar(t *testing.T) {
+	t.Setenv("SSH_AUTH_SOCK", "")
+	if SSHAgentAvailable() {
+		t.Error("SSHAgentAvailable() should be false when SSH_AUTH_SOCK is not set")
+	}
+}
